requirements.txt
python-docx
streamlit
PyPDF2
langchain-google-genai
python-dotenv
Pillow
langchain
reportlab
easyocr
pymupdf

